+++ default.nix
@@ -1,7 +1,4 @@
 { pkgs ? import <nixos-unstable> { } }:
-
-rec {
-
+{
   dibbler = pkgs.callPackage ./nix/dibbler.nix { };
-
 }
