+++ dibbler/conf.py
@@ -1,0 +1,6 @@
+# This module is supposed to act as a singleton and be filled
+# with config variables by cli.py
+
+import configparser
+
+config = configparser.ConfigParser()
