+++ dibbler/statistikkHelpers.py
@@ -1,11 +1,11 @@
 #! /usr/bin/env python
 # -*- coding: UTF-8 -*-
+
 import datetime
 from collections import defaultdict
-import operator
-from helpers import *
-import sys
-import db
+
+from .helpers import *
+from .models.db import *;
 
 def getUser():
 	while 1:
