+++ dibbler/text_based.py
@@ -5,26 +5,36 @@ import random
 import sys
 import traceback
 
-from helpers import *
-from text_interface.addstock import AddStockMenu
-from text_interface.buymenu import BuyMenu
-from text_interface.editing import *
-from text_interface.faq import FAQMenu
-from text_interface.helpermenus import Menu
-from text_interface.mainmenu import MainMenu
-from text_interface.miscmenus import ProductSearchMenu, TransferMenu, AdjustCreditMenu, UserListMenu, ShowUserMenu, \
-    ProductListMenu
-from text_interface.printermenu import PrintLabelMenu
-from text_interface.stats import *
+from .helpers import *
+from .menus.addstock import AddStockMenu
+from .menus.buymenu import BuyMenu
+from .menus.editing import *
+from .menus.faq import FAQMenu
+from .menus.helpermenus import Menu
+from .menus.mainmenu import MainMenu
+
+from .menus.miscmenus import (
+    ProductSearchMenu,
+    TransferMenu,
+    AdjustCreditMenu,
+    UserListMenu,
+    ShowUserMenu,
+    ProductListMenu,
+)
+
+from .menus.printermenu import PrintLabelMenu
+from .menus.stats import *
+
+from .conf import config
 
 random.seed()
 
 
-if __name__ == '__main__':
-    if not conf.stop_allowed:
+def main():
+    if not config.getboolean('general', 'stop_allowed'):
         signal.signal(signal.SIGQUIT, signal.SIG_IGN)
 
-    if not conf.stop_allowed:
+    if not config.getboolean('general', 'stop_allowed'):
         signal.signal(signal.SIGTSTP, signal.SIG_IGN)
 
     main = MainMenu('Dibbler main menu',
@@ -53,7 +63,7 @@ if __name__ == '__main__':
                            ],
                     exit_msg='happy happy joy joy',
                     exit_confirm_msg='Really quit Dibbler?')
-    if not conf.quit_allowed:
+    if not config.getboolean('general', 'quit_allowed'):
         main.exit_disallowed_msg = 'You can check out any time you like, but you can never leave.'
     while True:
         # noinspection PyBroadException
@@ -65,8 +75,12 @@ if __name__ == '__main__':
         except:
             print('Something went wrong.')
             print(f'{sys.exc_info()[0]}: {sys.exc_info()[1]}')
-            if conf.show_tracebacks:
+            if config.getboolean('general', 'show_tracebacks'):
                 traceback.print_tb(sys.exc_info()[2])
         else:
             break
         print('Restarting main menu.')
+
+
+if __name__ == '__main__':
+  main()
