+++ example-config.ini
@@ -1,0 +1,18 @@
+[general]
+quit_allowed = true
+stop_allowed = false
+show_tracebacks = true
+input_encoding = 'utf8'
+
+[database]
+url = postgresql://robertem@127.0.0.1/pvvvv
+
+[limits]
+low_credit_warning_limit = -100
+user_recent_transaction_limit = 100
+
+# See https://pypi.org/project/brother_ql/ for label types
+# Set rotate to False for endless labels
+[printer]
+label_type = "62"
+label_rotate = false
