--- special_barcodes/barcode.tex
@@ -1,47 +1,0 @@
-\documentclass[a4paper]{article}
-\usepackage[utf8]{inputenc}
-\usepackage{fullpage}
-\usepackage{pstricks}
-\usepackage{pst-barcode}
-
-\title{Dibbler Special Barcodes}
-\date{}
-\begin{document}
-
-\maketitle
-\setlength\tabcolsep{25pt}
-
-
-\begin{tabular}{ccc}
-\begin{pspicture}(3,1.2in) 
-\psbarcode[scalex=0.8,scaley=0.8]{4242424242420}{includetext guardwhitespace}{ean13}
-\end{pspicture} 
-&
-\begin{pspicture}(3,1.2in) 
-\psbarcode[scalex=0.8,scaley=0.8]{7640140330815}{includetext guardwhitespace}{ean13}
-\end{pspicture}
-&
-\begin{pspicture}(3,1.2in) 
-\psbarcode[scalex=0.8,scaley=0.8]{7024850087007}{includetext guardwhitespace}{ean13}
-\end{pspicture}
-\\
-PVV-skjorte &
-Nespresso-kapsel &
-Risengrynsgraut 
-\\
-\begin{pspicture}(3,1.2in) 
-\psbarcode[scalex=0.8,scaley=0.8]{7024850087014}{includetext guardwhitespace}{ean13}
-\end{pspicture}
-&
-\begin{pspicture}(3,1.2in) 
-\psbarcode[scalex=0.8,scaley=0.8]{5000159410946}{includetext guardwhitespace}{ean13}
-\end{pspicture}
-&
-\\
-Rømmegraut
-&
-Snickers mini
-&
-\end{tabular}
-
-\end{document}
