--- torje_conf.py
@@ -1,5 +1,0 @@
-db_url = 'postgresql://torjehoa_dibblerdummy:1234@postgres/torjehoa_dibblerdummy'
-quit_allowed = True
-stop_allowed = False
-show_tracebacks = True
-input_encoding = 'utf8'
